# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_ftp.BandwidthConfig:
    properties:
      download_kbps:
        description: 下载速率限制 (KB/s), 0 表示无限制
        type: integer
      enabled:
        type: boolean
      per_user:
        description: 是否按用户独立限制
        type: boolean
      upload_kbps:
        description: 上传速率限制 (KB/s), 0 表示无限制
        type: integer
    type: object
  internal_ftp.Config:
    properties:
      allow_anonymous:
        description: 允许匿名登录
        type: boolean
      bandwidth_limit:
        allOf:
        - $ref: '#/definitions/internal_ftp.BandwidthConfig'
        description: 带宽限制
      enabled:
        type: boolean
      max_connections:
        description: 最大连接数
        type: integer
      pasv_host:
        description: 被动模式对外 IP
        type: string
      pasv_port_end:
        description: 被动模式端口范围结束
        type: integer
      pasv_port_start:
        description: 被动模式端口范围起始
        type: integer
      port:
        description: FTP 端口 (默认 21)
        type: integer
      root_path:
        description: 根目录
        type: string
      virtual_dirs:
        additionalProperties:
          type: string
        description: '虚拟目录映射: 虚拟路径 -> 实际路径'
        type: object
    type: object
  internal_ftp.UpdateConfigRequest:
    properties:
      enabled:
        type: boolean
    type: object
...
  internal_web.GenericResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
...
  internal_web.VolumeCreateRequest:
    properties:
      devices:
        example:
        - /dev/sda
        - /dev/sdb
        items:
          type: string
        type: array
      name:
        example: data-vol
        type: string
      profile:
        description: single, raid0, raid1, raid10, raid5, raid6
        example: raid1
        type: string
    required:
    - devices
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nas-os.io
    name: NAS-OS 团队
    url: https://github.com/nas-os/nas-os
  description: |-
    NAS-OS 是一个现代化的网络存储操作系统，提供卷管理、用户管理、共享管理、网络配置等功能。

    ## 功能模块
    - **卷管理**: Btrfs 卷创建、快照、RAID 配置
    - **用户管理**: 用户/组管理、认证授权
    - **共享管理**: SMB/NFS 共享配置
    - **网络管理**: 网络接口、DDNS、防火墙、端口转发
    - **Docker 管理**: 容器、镜像、应用商店
    - **插件系统**: 可扩展的插件架构
    - **配额管理**: 存储配额控制
    - **性能监控**: 系统性能监控与报告
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NAS-OS API
...
paths:
...
  /ftp/config:
    get:
      consumes:
      - application/json
      description: 获取 FTP 服务器配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取 FTP 配置
      tags:
      - ftp
    put:
      consumes:
      - application/json
      description: 更新 FTP 服务器配置
      parameters:
      - description: 配置
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/internal_ftp.Config'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 更新 FTP 配置
      tags:
      - ftp
...
  /ftp/transfers:
    delete:
      consumes:
      - application/json
      description: 清除所有 FTP 传输日志
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 清除传输日志
      tags:
      - ftp
    get:
      consumes:
      - application/json
      description: 获取 FTP 传输日志列表，支持过滤
      parameters:
      - description: 用户名过滤
        in: query
        name: username
        type: string
      - description: 传输方向 (upload/download)
        in: query
        name: direction
        type: string
      - description: 是否成功 (true/false/all)
        in: query
        name: success
        type: string
      - description: 开始时间 (RFC3339)
        in: query
        name: start_time
        type: string
      - description: 结束时间 (RFC3339)
        in: query
        name: end_time
        type: string
      - description: 返回数量限制
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取传输日志列表
      tags:
      - ftp
  /ftp/transfers/config:
    get:
      consumes:
      - application/json
      description: 获取传输日志配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取日志配置
      tags:
      - ftp
    put:
      consumes:
      - application/json
      description: 更新传输日志配置
      parameters:
      - description: 配置
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/internal_ftp.UpdateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 更新日志配置
      tags:
      - ftp
  /ftp/transfers/stats:
    get:
      consumes:
      - application/json
      description: 获取 FTP 传输统计信息
      parameters:
      - description: 统计周期 (1h/24h/7d/30d)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取传输统计
      tags:
      - ftp
...
  /volumes:
    get:
      consumes:
      - application/json
      description: 获取系统中所有 Btrfs 卷的列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
      summary: 列出所有卷
      tags:
      - volumes
    post:
      consumes:
      - application/json
      description: 使用指定设备和配置创建新的 Btrfs 卷
      parameters:
      - description: 卷创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_web.VolumeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
      summary: 创建新卷
      tags:
      - volumes
  /volumes/{name}:
    delete:
      consumes:
      - application/json
      description: 删除指定的 Btrfs 卷
      parameters:
      - description: 卷名称
        in: path
        name: name
        required: true
        type: string
      - description: 强制删除
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
      summary: 删除卷
      tags:
      - volumes
    get:
      consumes:
      - application/json
      description: 根据卷名称获取卷的详细信息
      parameters:
      - description: 卷名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
        "404":
          description: 卷不存在
          schema:
            $ref: '#/definitions/internal_web.GenericResponse'
      summary: 获取卷详情
      tags:
      - volumes
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API 密钥认证
    in: query
    name: token
    type: apiKey
  BearerAuth:
    description: JWT 认证令牌，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...